Adds Playtomic scraper (not ready) and first functions of new search aproach
The new search aproach searches for all possible time_blocks in the date. It returns much more results to Bubble, but is helpful because the user won't need another scraping phase (slow) to change some of the search parameters. Bubble will filter the results acording to the parameters.

Other mayor fix is that al dates and times are now treated as datetime objects, which will fix issues regarding border conditions that appear when block's final time is later than midnight